+++ tests/test_rhg_compute_tools.py
@@ -3,9 +3,18 @@
 
 """Tests for `rhg_compute_tools` package."""
 
+from time import sleep
+
 import pytest
+from dask_gateway import Gateway
 
 from rhg_compute_tools import kubernetes
+
+
+class LocalGateway(Gateway):
+    def __init__(self):
+        # initialize local gateway
+        super().__init__("http://127.0.0.1:8000")
 
 
 def monkeypatch_cluster(func):
@@ -16,9 +25,7 @@ def monkeypatch_cluster(func):
         def mock_dask_Client(cluster):
             return cluster
 
-        monkeypatch.setattr(
-            "dask_kubernetes.KubeCluster.from_dict", mock_KubeCluster_from_dict
-        )
+        monkeypatch.setattr("dask_gateway.Gateway", LocalGateway)
 
         monkeypatch.setattr("dask.distributed.Client", mock_dask_Client)
 
@@ -29,47 +36,34 @@ def monkeypatch_cluster(func):
 
 @pytest.mark.parametrize("mem,cpu,scale", [(None, None, None)])
 @monkeypatch_cluster
-def test_create_worker(mem=None, cpu=None, scale=None):
+def test_create_worker_gateway(mem=None, cpu=None, scale=None):
 
-    cluster, client = kubernetes.get_cluster(
-        template_path="tests/resources/worker_template.yml"
+    # can only test options that are valid on a UnsafeLocalBackend (e.g. nothing that
+    # has to do with kubernetes pod config)
+    client, cluster = kubernetes._get_cluster_dask_gateway(
+        profile="micro",
+        cred_path="path/to/test_cred.json",
+        env_items={"TEST_ITEM": "TEST_RESULT"},
     )
 
-    mem = cluster["spec"]["containers"][0]["args"][5]
-    assert mem == "11.50GB", mem
+    cluster.scale(1)
+    sleep(2)
 
-    res_lim = cluster["spec"]["containers"][0]["resources"]["limits"]
-    assert res_lim["memory"] == "11.50G", res_lim["memory"]
-    assert res_lim["cpu"] == "1.75", res_lim["cpu"]
+    # check env_items worked
+    def return_test_env_var(key):
+        import os
 
-    res_req = cluster["spec"]["containers"][0]["resources"]["requests"]
-    assert res_req["memory"] == "11.50G", res_req["memory"]
-    assert res_req["cpu"] == "1.75", res_req["cpu"]
+        return os.environ[key]
+
+    assert client.gather(
+        client.map(return_test_env_var, ["GOOGLE_APPLICATION_CREDENTIALS", "TEST_ITEM"])
+    ) == ["/opt/gcsfuse_tokens/test_cred.json", "TEST_RESULT"]
+
+    # scale down gracefully
+    cluster.scale(0), cluster.close(), client.close()
 
 
 size_test_params = [(35, 7, None), (1, 6, None), (4, 1, None)]
-
-
-@pytest.mark.parametrize("mem,cpu,scale", size_test_params)
-@monkeypatch_cluster
-def test_size_worker(mem, cpu, scale):
-
-    cluster, client = kubernetes.get_cluster(
-        template_path="tests/resources/worker_template.yml", memory_gb=mem, cpus=cpu
-    )
-
-    mem_arg = cluster["spec"]["containers"][0]["args"][5]
-    assert abs(mem - float(mem_arg.strip("GB"))) < 0.01, mem_arg
-
-    res_lim = cluster["spec"]["containers"][0]["resources"]["limits"]
-    mem_size = float(res_lim["memory"].strip("G"))
-    assert abs(mem_size - mem) < 0.01, res_lim["memory"]
-    assert abs(float(res_lim["cpu"]) - cpu) < 0.01, res_lim["cpu"]
-
-    res_req = cluster["spec"]["containers"][0]["resources"]["requests"]
-    mem_size = float(res_req["memory"].strip("G"))
-    assert abs(mem_size - mem) < 0.01, res_req["memory"]
-    assert abs(float(res_req["cpu"]) - cpu) < 0.01, res_req["cpu"]
 
 
 scale_test_params = [
@@ -78,25 +72,3 @@ scale_test_params = [
     (None, None, 2),
     (None, None, 4),
 ]
-
-
-@pytest.mark.parametrize("mem,cpu,scale", scale_test_params)
-@monkeypatch_cluster
-def test_scale_worker(mem, cpu, scale):
-
-    cluster, client = kubernetes.get_cluster(
-        template_path="tests/resources/worker_template.yml", scaling_factor=scale
-    )
-
-    mem_arg = cluster["spec"]["containers"][0]["args"][5]
-    assert abs((scale * 11.5) - float(mem_arg.strip("GB"))) < 0.01, mem_arg
-
-    res_lim = cluster["spec"]["containers"][0]["resources"]["limits"]
-    mem_val = float(res_lim["memory"].strip("G"))
-    assert abs(mem_val - (scale * 11.5)) < 0.01, res_lim["memory"]
-    assert abs(float(res_lim["cpu"]) - (scale * 1.75)) < 0.01, res_lim["cpu"]
-
-    res_req = cluster["spec"]["containers"][0]["resources"]["requests"]
-    mem_val = float(res_req["memory"].strip("G"))
-    assert abs(mem_val - (scale * 11.5)) < 0.01, res_req["memory"]
-    assert abs(float(res_req["cpu"]) - (scale * 1.75)) < 0.01, res_req["cpu"]